requirements.txt
databases==0.7.0
python-dotenv==1.0.0
aiosqlite==0.18.0
python-jose==3.3.0
python-multipart==0.0.6
fastapi==0.95.0
fastapi-sso==0.6.4
uvicorn==0.21.1
